+++ composite/composite/CEllipse.h
@@ -5,15 +5,13 @@
 class CEllipse : public CShape
 {
 public:
-	CEllipse(PointD const& center, float horizontalRadius, float verticalRadius);
+	CEllipse(PointD const& center, double horizontalRadius, double verticalRadius);
 	RectD GetFrame() override;
-	void SetFrame(const RectD & rect) override;
-	void DrawFigure(ICanvas & canvas)const override;
+	void SetFrame(const RectD& rect) override;
+	void DrawFigure(ICanvas& canvas) const override;
 
 private:
 	PointD m_center;
-	float m_horizontalRadius;
-	float m_verticalRadius;
-
+	double m_horizontalRadius;
+	double m_verticalRadius;
 };
-
